[Chore] Extract FoodieAppBar and FoodieSidebarMenu
Extract `FoodieAppBar` and `FoodieSidebarMenu` components, as a refactor
measure inorder to ensure we conform to the Single Responsibility
Principle.
